+++ packages/mix/lib/src/widgets/pressable_widget.dart
@@ -4,6 +4,7 @@ import '../core/factory/style_mix.dart';
 import '../core/widget_state/internal/gesture_mix_state.dart';
 import '../core/widget_state/internal/interactive_mix_state.dart';
 import '../core/widget_state/internal/listener_mix_state.dart';
+import '../core/widget_state/internal/mix_widget_state_builder.dart';
 import '../core/widget_state/widget_state_controller.dart';
 import '../internal/constants.dart';
 import '../specs/box/box_widget.dart';
@@ -197,7 +198,7 @@ class PressableWidgetState extends State<Pressable> {
       excludeFromSemantics: widget.excludeFromSemantics,
       hitTestBehavior: widget.hitTestBehavior,
       unpressDelay: widget.unpressDelay,
-      child: ListenerMixStateWidget(
+      child: PointerListenerMixStateWidget(
         child: InteractiveMixStateWidget(
           enabled: widget.enabled,
           onFocusChange: widget.onFocusChange,
@@ -300,25 +301,23 @@ class _InteractableState extends State<Interactable> {
 
   @override
   Widget build(BuildContext context) {
-    return ListenerMixStateWidget(
-      child: InteractiveMixStateWidget(
-        enabled: widget.enabled,
-        onFocusChange: widget.onFocusChange,
-        autofocus: widget.autofocus,
-        focusNode: widget.focusNode,
-        onKey: widget.onKey,
-        onShowFocusHighlight: widget.onShowFocusHighlight,
-        onShowHoverHighlight: widget.onShowHoverHighlight,
-        onKeyEvent: widget.onKeyEvent,
-        canRequestFocus: widget.canRequestFocus,
-        mouseCursor: widget.mouseCursor,
-        shortcuts: widget.shortcuts,
+    return InteractiveMixStateWidget(
+      enabled: widget.enabled,
+      onFocusChange: widget.onFocusChange,
+      autofocus: widget.autofocus,
+      focusNode: widget.focusNode,
+      onKey: widget.onKey,
+      onShowFocusHighlight: widget.onShowFocusHighlight,
+      onShowHoverHighlight: widget.onShowHoverHighlight,
+      onKeyEvent: widget.onKeyEvent,
+      canRequestFocus: widget.canRequestFocus,
+      mouseCursor: widget.mouseCursor,
+      shortcuts: widget.shortcuts,
+      controller: _controller,
+      actions: widget.actions,
+      child: MixWidgetStateBuilder(
         controller: _controller,
-        actions: widget.actions,
-        child: MixWidgetStateBuilder(
-          controller: _controller,
-          builder: (context) => widget.child,
-        ),
+        builder: (context) => widget.child,
       ),
     );
   }
