+++ packages/mix/test/src/core/mix_state/internal/gesture_mix_state_test.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mix/src/core/widget_state/internal/gesture_mix_state.dart';
+import 'package:mix/src/core/widget_state/internal/mix_widget_state_builder.dart';
 import 'package:mix/src/core/widget_state/widget_state_controller.dart';
 
 void main() {
