+++ packages/mix/test/src/core/mix_state/internal/interactive_mix_state_test.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mix/src/core/widget_state/internal/mix_widget_state_builder.dart';
 import 'package:mix/src/core/widget_state/widget_state_controller.dart';
 
 void main() {
