+++ packages/mix/test/src/core/widget_state/internal/listener_mix_state_test.dart
@@ -8,7 +8,7 @@ void main() {
     testWidgets('updates pointer position on hover',
         (WidgetTester tester) async {
       await tester.pumpWidget(
-        const ListenerMixStateWidget(
+        const PointerListenerMixStateWidget(
           child: SizedBox(width: 100, height: 100),
         ),
       );
@@ -17,39 +17,19 @@ void main() {
       await gesture.addPointer(location: Offset.zero);
       addTearDown(gesture.removePointer);
 
-      final beforePosition =
-          ListerMixStateProvider.of(tester.element(find.byType(SizedBox)))!
-              .pointerPosition;
+      final beforePosition = PointerListenerMixWidgetState.of(
+              tester.element(find.byType(SizedBox)))!
+          .pointerPosition;
 
       await gesture.moveTo(const Offset(50, 50));
       await tester.pumpAndSettle();
 
-      final afterPosition =
-          ListerMixStateProvider.of(tester.element(find.byType(SizedBox)))!
-              .pointerPosition;
+      final afterPosition = PointerListenerMixWidgetState.of(
+              tester.element(find.byType(SizedBox)))!
+          .pointerPosition;
       expect(beforePosition, isNull);
 
       expect(afterPosition!.offset, equals(const Offset(50, 50)));
-    });
-
-    testWidgets('calls onHover callback', (WidgetTester tester) async {
-      PointerHoverEvent? hoverEvent;
-      await tester.pumpWidget(
-        ListenerMixStateWidget(
-          onHover: (event) => hoverEvent = event,
-          child: const SizedBox(width: 100, height: 100),
-        ),
-      );
-
-      final gesture = await tester.createGesture(kind: PointerDeviceKind.mouse);
-      await gesture.addPointer(location: Offset.zero);
-      addTearDown(gesture.removePointer);
-
-      await gesture.moveTo(const Offset(50, 50));
-      await tester.pumpAndSettle();
-
-      expect(hoverEvent, isNotNull);
-      expect(hoverEvent!.localPosition, equals(const Offset(50, 50)));
     });
   });
 }
