+++ packages/mix/test/src/widgets/pressable/pressable_widget_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mix/mix.dart';
-import 'package:mix/src/core/widget_state/widget_state_controller.dart';
+import 'package:mix/src/core/widget_state/internal/mix_widget_state_builder.dart';
 
 import '../../../helpers/testing_utils.dart';
